feat: error handling

+++ backend/src/authors/authors.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { CreateAuthorDto } from './dto/create-author.dto';
 import { UpdateAuthorDto } from './dto/update-author.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { AuthorNotFoundException } from './exceptions/author-not-found';
 
 @Injectable()
 export class AuthorsService {
@@ -17,14 +18,23 @@ export class AuthorsService {
     return this.prisma.author.findMany();
   }
 
-  findOne(id: number) {
-    return this.prisma.author.findUnique({
+  async findOne(id: number) {
+    const author = await this.prisma.author.findUnique({
       where: { id: id },
       include: { books: true },
     });
+    if (!author) {
+      throw new AuthorNotFoundException();
+    }
+    return author;
   }
 
-  update(id: number, updateAuthorDto: UpdateAuthorDto) {
+  async update(id: number, updateAuthorDto: UpdateAuthorDto) {
+    const existing = await this.prisma.author.findUnique({ where: { id } });
+    if (!existing) {
+      throw new AuthorNotFoundException();
+    }
+
     return this.prisma.author.update({
       data: updateAuthorDto,
       where: {
@@ -33,7 +43,12 @@ export class AuthorsService {
     });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const existing = await this.prisma.author.findUnique({ where: { id } });
+    if (!existing) {
+      throw new AuthorNotFoundException();
+    }
+
     return this.prisma.author.delete({
       where: { id: id },
     });
